requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
asgiref==3.11.1
    # via
    #   django
    #   django-cors-headers
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
build==1.4.4
    # via pip-tools
certifi==2026.4.22
    # via requests
charset-normalizer==3.4.7
    # via requests
click==8.3.3
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
    #   pytest
coverage==7.13.5
    # via -r requirements/dev.in
django==6.0.4
    # via
    #   -r requirements/base.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-cors-headers==4.9.0
    # via -r requirements/base.in
django-filter==25.2
    # via -r requirements/base.in
djangorestframework==3.17.1
    # via
    #   -r requirements/base.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via -r requirements/base.in
drf-spectacular==0.29.0
    # via -r requirements/base.in
idna==3.13
    # via requests
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.3.0
    # via pytest
jsonschema==4.26.0
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
packaging==26.1
    # via
    #   build
    #   pytest
    #   wheel
pillow==12.2.0
    # via -r requirements/base.in
pip-tools==7.5.3
    # via -r requirements/dev.in
pluggy==1.6.0
    # via pytest
psycopg2-binary==2.9.12
    # via -r requirements/dev.in
pygments==2.20.0
    # via pytest
pyjwt==2.12.1
    # via djangorestframework-simplejwt
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==9.0.3
    # via
    #   -r requirements/dev.in
    #   pytest-django
pytest-django==4.12.0
    # via -r requirements/dev.in
python-dotenv==1.2.2
    # via -r requirements/base.in
pyyaml==6.0.3
    # via drf-spectacular
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.33.1
    # via -r requirements/base.in
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.5
    # via django
tzdata==2026.1
    # via django
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.6.3
    # via requests
wheel==0.47.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
